End of section 18

+++ DatingApp-SPA/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/Message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -10,14 +11,14 @@ import { MessageService } from '../_services/message.service';
 })
 export class MessagesComponent implements OnInit {
 
-  messages: Message[]=[];
+  messages: Message[] = [];
   pagination!: Pagination;
   container='Unread';
   pageNumber=1;
   pageSize=5;
   loading=false;
 
-  constructor(private messageService:MessageService) { }
+  constructor(private messageService:MessageService,private confirmService:ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -33,9 +34,14 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id:number){
-    this.messageService.deleteMessage(id).subscribe(()=>{
-      this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
+    this.confirmService.confirm('Confirm delete message.','This cannot be undone!').subscribe(result=>{
+      if(result){
+        this.messageService.deleteMessage(id).subscribe(()=>{
+          this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
+        });
+      }
     })
+   
   }
 
   pageChanged(event:any){
